requirements.txt
fastapi==0.115.0
openai==1.51.0
qdrant-client==1.7.3
fastembed>=0.8.0
python-dotenv==1.0.1
uvicorn==0.30.6
httpx==0.27.0
motor==3.6.0
slowapi==0.1.9
pytest==8.3.3
langdetect==1.0.9
